refactor: remove email and rewrite update user name function
