+++ src/utils/storage.ts
@@ -1,3 +1,6 @@
+import { AxiosError } from "axios";
+import getUserInfo from "../API/getUserInfo";
+
 export default function setRegister(
   token: string,
   name: string,
@@ -11,4 +14,12 @@ export default function setRegister(
       isAuth,
     })
   );
+}
+
+export function updateAuthRegister(token: string, isAuth: boolean) {
+  try {
+    getUserInfo(token).then((value) => setRegister(token, value.name, isAuth));
+  } catch (error) {
+    throw new Error((error as AxiosError).message);
+  }
 }
